small clarification/comment
Got confused looking at this code for a bit forgot what was implied by center_x/y, better updated it and added a comment that this is just a temporary variable for moving the video object based on the top left of the video_item. Especially because this does not refer to the self.center_x which refers to the center of the video_item itself (video width/2).

I feel that the whole VideoBackgroundManager needs a reformat to make it a lot easier to read and simplify calculations.
Will work on that after I overcome this flu I picked up....