add a few initial variable setting in get_video_aspect_ratio
For video objects, the goal is to make the initial get_video_aspect_ratio() set as many variables and do as many calculations as it can, such that it only runs all of these once. The current problem is that even if the video itself doesn't change, whenever init_center_point() or a few more functions are called for the videoBackgroundManager, it does some calculations which shouldn't need a change.
This commit removes a few instancing of variables from init_center_point() which is called every time the window is resized, and places them in get_video_aspect_ratio() which is only ran on start, or when a new video is loaded (load_new_video).

There are still quite a few other instancing/calculations which can be moved to get_video_aspect_ratio and a bunch of simplifications to do.

Still recovering from covid, work will speed up as things get better.