small comment upgrade.
Small description change for self.center_x/y in VideoBackgroundManager.

Basically self.center_x/y are coordinates for the focal point upon which zooms are applied. With debug enabled the focal point should always be the red dot, but it is stored as a reference coordinate inside the video_item (for instance no offset changes the coordinate = video_width/2, video_height/2).