Keybind restore no longer minimizes before restoring
Originally had to set it up like this for some nonsensical reason. Believe that this was a problem before the engine was rewritten, which should have solved a lot of problems with the window sizing/placing/keeping track of the window state.

Didn't recognize any problems with this implementation, it is possible that the special case which warranted the minimizing before maximizing was esoteric enough to avoid in this small testing however. so will keep testing to make sure the keybind restore is working as expected.