Fix swapping icons not updating icon if paths matched.
Fixed an oversight that would keep the cached icon image on swapping the files, if the two icons shared the same directory and file name. For instance, icon2.png when swapped would match the same file path as the existing one, thus it would not try and load the new image because it already had an icon cached with the same file path.