Simplified ImageBgMngr caluculations
Now when the image is loaded it keeps track of the pixmap_width and height. Then removed calculations done on every load and every move_background() call and replaced it with the width/height reference.

Very small performance increase, if at all noticeable, more for readability and standardization than performance.