Added settings.py: font, font_size, border_color, border_width
Will be a while before these can actually be edited but made these settings for a refactoring of DesktopIcons class soon.

I.e. these are stored in the settings.json but do not have a way of setting them in the program yet. Making these settings is mostly just for accessing them from a future desktop_icon.py class file.