Removed non-text overrides for Dark themes
Somehow originally overlooked setting primaryTextColor when I first added themes, and did the weird workaround of setting secondaryColor and secondaryDarkColor to make the text readable in the LineEdit and drop down menus.

Since it really was as simple as just setting primaryTextColor to white, I have reverted the explicit setting of secondaryColor and secondaryDarkColor.

Now the only color that gets overriden for dark is the primaryTextColor to white.