Fixed variable init happening after use.
Fixed a problem where it was applying theme (which sets all the theme color variables) BEFORE initiation of the class variables. Thus it was creating them with the correct color values, then running the init which would overwrite the variables back to None.

This caused a bug where when launching in dark mode the color value received for the background was None thus would set the color background to #000000 instead of the theme's color.