Fix: stretched appearance for tabs in settings menu
Fixed the appearance inside the "icon" tab in settings menu to essentially force the separator and setting/slider to the top.

Also noticed this problem in "General" tab if I were to remove a few settings. Thus you would expect a gap at the bottom but instead it would mess up the entire spacing of the tab stretching out the vertical padding between lines to fill the tab.
So it has been added to "General" tab as well for consistency, since multiple settings from the "General" tab will be moved to the "Icons" tab shortly.