Added some clarification to the Default font size setting
Updated "Font Size" setting to "Default Font size" and a tooltip to emphasize that it may be overridden by each icon, but serves as the default size.

Changed my mind on how the Default Font size will work, so will need to rework some of the way it saves it to desktop.py.

Basically it should follow the "Default Font size" setting UNLESS the user has changed the font size. Basically the "font_size" icon setting needs to default to "default" or something and then only save for instance "10" when the user actually edited the font size. Then use the "default" to signify that when setting the font size it should just use the Default Font size setting.