Font size now defaults to Default Font size.
Plan going forward is to make sure that unless a font is specifically changed inside the desktop_grid_menu (icon edit menu) it should maintain a font_size of -1. When it has -1 value it then instead of creating a font with -1 size it loads from the Main settings Default Font size."

Therefore if a user never goes into an icon and changes the font manually it will always use the Default Font setting in the settings menu.