Added Appearance tab to icon edit menu.
Added a tab and auto loads in either the overridden font size (if previously changed from default OR the Default Font size from main settings menu).

Also consolidated a check for if default font size should be used or if it is overridden into the function get_icon_font_size(row, col) in config.py. This effectively does the check and returns the proper font size that the icon should use.