Some cleanup.
Some cleanup/rename of desktop_grid_menu.py I did while looking through and planning how to solve a problem I noticed.

Problem was differentiating when an icon has an overridden font_size. It is easy to tell when you set it initially but there unless I have a way to tell when it is already saved as a custom font_size it can create a bug where a custom icon font size is changed back to default because the last time it was saved the font_size for the icon matched the Default font size.

So plan is to just add another config boolean for "Uses Default Font size" to easily differentiate if the icon is already using a default font size. Which will allow accurate checking of the icons font size status.