Icons now properly track "use_default_font_size"
Added a logic check for saving True/False for "use_default_font_size"

In order to save as true the icon grid_menu must be launched with "use_default_font_size" already true. Then either the font_size Spinbox setting must not be touched OR the value must be returned to the value when the grid_menu was opened in order to keep "use_default_font_size" True. Otherwise it becomes False. (Either was False already, or font_size was changed to another value).