Added "Reset" functionality for font_size
Now when clicking the Reset button in an icon's edit menu it will reset the "font size" to the default settings "font_size".

If then the user does not change the font size it will reset the size and the "use_default_font_size" setting to True.

Still need to implement a visual update to resize the font on settings change, as of right not it only updates the font sizes when you change the Settings Menu "font size"