Refactor: desktop_grid_menu -> icon_edit_menu.py
Refactored: "desktop_grid_menu.py" to "icon_edit_menu.py"

Main reason for this is that it is a much better descriptor for what the class actually is. Referring to settings that were in the "desktop_grid_menu" got confusing, between the default settings menu and the "desktop_grid_menu".
It should be much easier to differentiate  now by referring to either the "Settings menu" or the "icon_edit_menu" settings.