edit_mode/normal_mode -> draw_red_border/hide_red_border
Naming change from edit_mode_icon() and normal_mode_icon() to draw_red_border() and hide_red_border() respectively.

Naming change is mostly just to avoid any confusion with an "edit_mode" of an icon and the name better correlates to the RedBorderItem which it either draws or hides, hence the name change.