Fixed closing editing with temp icon not removing icon.
Forgot to add the remove_temp_icon() to the closeEvent.

Also added it just as a fallback to the end of the contextMenuEvent in desktop_grid.py. As all temp icons should only ever exist while still in the contextMenuEvent, removing it upon closing the contextMenu should remove catch any previously uncaught temp_icons.