Fixed icons not previewing on changing icon_path.
Fixed a problem where editing the icon_path in the icon_edit_menu was not actively previewing the current path but the cached path. Thus even on an invalid icon_path it would preview the already saved icon_path. This has been updated to now reload the pixmap every time it is changed to properly preview the current icon_path.

This is probably a problem created when it was swapped to a cached system for performance back in V0.3.000.