Fixed edit icon menu appearing partially below screen
Fixed a problem where if the global position of the window was negative, the code to check if the menu would appear below the bottom of the screen would not trigger, thus rendering part of the window below the taskbar/off screen.

For reference this would occur when a monitor is vertically above the main monitor in windows.