added "use_default_font_color" setting for DesktopIcon
Added a "use_default_font_color" setting in config for keeping track if the DesktopIcon uses the default font color or not.

Functionality will pretty much mirror the current functionality of the "use_default_font_size" at the moment.

It will default to true.
If saved with a different font color, will become false.
If false it will load the set font_color when loaded/reloaded.
In settings if you press reset to reset all font colors to a color, it will reset all DesktopIcons to that color.

Eventually will probably add another setting like this for font type, allowing users to set different icons to use different fonts.

Also need to get around to fixing the way that the label area is allocated and how the multi-line text works.