Frontend for changing font color for DesktopIcon
Added a line "Font Color" to the DesktopIcon appearance tab of the Edit Icon menu.

The line now contains the label "Font Color" and a button which previews the self.font_color Upon clicking the button it opens the ColorDialog which lets the user select a color.

As this is just the frontend it does not save the color. It will be updated to save soon.