Added font_color_changed
Added font_color_changed to keep track of if the user actually changed the color. Only changes to true if the user saves the color dialog with a different color than when it was opened. I.e. if the user opens it but doesn't change anything or changes, then changes back do not count it as being overridden and changed.

Still does not currently save the use_default_font_color. Few more steps and it will be implemented after some testing.