Fixed typo: use_default_font_size to always return true
Fixed a typo in get_icon_data() which used a misnamed setting in the item.get() call thus would always return the fallback True.

This seemingly mostly affected actually just doing the debugging to see it's value, as it was still working perfectly fine, the only case it seemed to be wrong was ironically the debugging (right click logging statements). It was saved correctly in the file and when actually referenced from config was loaded via "get_icon_font_size" so it was overlooked.