Previewing icon "font color"
When changing the color in a desktopIcon's appearance tab it now previews the font color on the icon.

This still is set to always save with the default font color. I am working on the tweaks which could possibly cause annoyances first then will test and deploy a finished version which will save both the color and the "use_global_font_color" setting soon.