icon_edit_menu load_data() and bugfix.
Separated out the block which loads all the data from the icon/global fields into it's own function load_data().

It also now properly loads the font size and font color when creating a brand new (not in desktop.json) desktopIcon. Previously it was only loading the size and default color if the icon already existed in the desktop.json.

Also fixed a problem which prevented saving desktopIcons due to a typo calling a child's function (which did not exist) instead of the class function.

Custom font colors still not implemented. Still fixing and rewriting some stuff I came across which desperately needed it.