Deleting an icon now deletes from the desktop.json
Originally implemented it to work like this, but removed it to just set it to default. Thought process was basically just thinking it was less likely to cause an error. I think it can be pretty reliably done now without causing an error due to some updating with the config.py recently, and a better implementation for deleting the entry.

Thus it now fully deletes an entry from desktop.json if the user confirms to delete the desktop_icon.