Fixed typo.
Originally debated on changing the icon's config font_color setting to icon_font_color. Thought I reverted this before a change done yesterday or the day before, turns out I didn't and this made it to main branch....

Name has been reverted and it now properly displays the icons color.

Now there is only 1 small bug to fix, and that is that something is messing up with loading the initial color into an icon's appearance tab's "font color" setting where it always defaults to the global_font_colors color. Such that when saved without changing it, will revert it back to the global_font_color.