Added "reset" button to "Name font color" setting
Added a button "Reset" to the "Name font color" setting in the settings menu.

For now it just displays the font_size() warning and when clicking OK it does nothing. This will be updated to work as expected soon.

Also reformatted the add_icon_tab() to fit better logically with ordering of items