Normalized SettingsMenu closure.
close_settings is now always called upon closing the settings menu.

Inside close_settings() it now always calls redraw_all_icons() which keeps the icons up to date on every closing of the settings menu.

I was originally going to just make calls to self.parent().grid_widget.redraw_all_icons() everywhere it was going to close the settingsMenu. However I instead normalized closure for 2 reasons.
1. Easier to understand the flow and catch errors.
2. Placing the call before self.accept() will add a small delay before the SettingsMenu closes for the user.

It is better UI design for it to be a quicker close, and better
general design to have a normalized closure.