Slight adjustments to display_warning messagebox
Made it so that the minimum width can be a bit skinnier.

Also I removed a part which I believe I was originally trying to remove all other objects/widgets inside the basic QMessageBox but never got around to fully deleting all of them. And this also messed up the formatting of the messagebox and it looks better without removing these widgets so it has been removed for now.