Fixed video item "jumping" on load.
The video item now waits to show until after its transform has been updated. This creates a much smoother video background display on launch/swapping to video background.

Previously it would load the video, then perform the transforms thus jarringly "jumping" it into place.

As far as Image backgrounds go I am pretty sure that the image background is loaded first then moved like this video was but I believe that the image background is just so optimized that it is nearly impossible to have or notice it "jump" into place. Might look into doing it just in case a slow machine or a massive desktop background could possibly lead to it jumping.