Beginning of a shelf
Mostly scuffed extreme work in progress, For now this is just creating a rectangle that will be on the right side of the screen.

It will also only appear in "devbug" or "debug" mode for now as it might cause some problems and currently is drawing an outline around itself.