Shelf button (only appears in debug mode)
Added a QGraphicsWidget Shelf class that will hold all the shelf components.

For now is just a QButton that appears on the right side of the screen. but will control the extension of the shelf object with things like a settings menu and other stuff.

It will also be hooked up to only appear visible when the mouse is on within the area of the screen for the ShelfHoverItem, such that the button does not always exist there but shows itself when it is needed.