Shelf button now only appears if user mouses over ShelfHoverItem
Shelf extend button now only shows when the user hovers over the right side of the screen (area defined by ShelfHoverItem).

Also moved the ShelfHoverItem to the shelf.py file as it fits in well there.

Reminder the shelf currently only appears in "debug" mode (run by doing "py alternativedesktop.py debug" or "devbug" in a work environment (debug does not log to console, only to file, devbug logs to console).

Still need to write the code for extending the shelf. As of right now it only prints "toggle shelf called" when the button is clicked.