Stop clicking on Shelf components from triggering grid clicks.
When clicking on the shelf's toggle button, if double clicked it would trigger a double click on the desktop at that same location, would not be the biggest deal since the button appears on the far right and thus would 99.9% of cases never result in calling the mouseDoubleClickEvent. However when extended a double click to close the shelf might cause it to execute the icon directly underneath the button.