If shelf showing do not hide shelf. And added close_shelf()
When the shelf is open do not let show_button() hide the shelf button.

Secondly added a close_shelf() function which will actually be used to close the shelf even if the shelf is open.

Reminder these changes are only visible in "debug" or "devbug" mode. (py alternativedesktop.py devbug) etc.