Fixed attempt to remove unadded item from scene.
Fixed a problem where clicking on an icon without moving or dragging would cause a QT warning in console (was trying to remove an item from the scene, but the item had not yet been added).

It now only attempts to remove the item from the scene if the item has actually been added (uses visibility to track if it has been added).