First attempt at shelf sliding/content
Added actual content to the shelf such that when the button is pressed it slowly slides out and displays the inner content of the shelf. Then when the toggle shelf button is clicked again it closes the shelf (Although for now it immediately hides the content and doesnt play the animation).

Known issues:
Shelf moves itself to 0 on the y axis (very top of GraphicsView) when opened/closed.
Animation does not play for closing (immediately hides button/shelf).