Prep shelf for settings button.
Changed the sample button to "Settings Menu" button. and removed the test label from the shelf.

I think I will rewrite how the SettingsMenu is managed as it is currently managed by the desktop.py which makes referencing and opening it extremely annoying from the child desktop_grid.py or from shelf.py which has no relationship to easily call for showing the settings menu.