Removed base settings button
Removed the Settings button at the bottom of the desktop.

You can now access the settings menu by right click -> Settings menu on any empty desktop item or through a button in the shelf (hover over the right side of the screen, then open shelf by clicking the button that appears).

Removed the "debug/devbug" requirement for shelf. It now always appears, although it still shows the outline of the hover_item (will be hidden eventually).