Optimized config.py
Optimized config.py to use a ITEM_LOOKUP_TABLE to directly reference and access the JSON copy of the desktop.json file.

This allows for faster lookups for items.
Previously it would loop through all items in the JSON until it found one that matches the (row, col). Now it just checks the ITEM_LOOKUP_TABLE which would give it a direct reference to the item (if it exists).

Basically reduced the time complexity of many functions in config.py to O(1) from O(n). While n is still expected to be small (<100 total desktop icons expected). It is still a likely noticeable improvement.