Modularized closing/opening shelf.
Added open_shelf() and updated close_shelf() to consistently open/close the shelf in a standardized way. toggle_shelf() now directly calls these or they can be called from elsewhere (like the instance for mousePressEvent in desktop_grid which calls to close_shelf() if shelf is open.

This fixes the inconsistency with the mousePressEvent on the desktop_grid where it would close instantly to now close as normal and avoid getting stuck open.