ShelfHoverItem now uses Screen% for sizing
ShelfHoverItem(width, height, ...) Now uses width and height as a % of the screen. For instance width = 10 sets the width of the hoverItem = 10% of width of screen. Same with height where height = 70 sets the height of the hoverItem to 70% of the screen height.

These adapt and update for every resize of the window to maintain the percentage based size.